fix(jsii): deprecation warnings use this.constructor before calling super (#3528)
Instead of using `this.constructor`, inline the class name (which has
the same value as `this.constructor` anyway). This removes what can be
seen as a semantic issue (as it is usually illegal to refer to `this`
before having called `super` in strict mode).

Fixes #3527 

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0