fix(python): cannot import nested submodules
Nested submodules were not registered in the `__all__` list of their
parent, and so were not visible to import statement such as
`from fully.qualified.submodule.path import Name`.

This changes how submodules get loaded so that they are always declared
and loaded with the context of their immediate parent sub•module, which
guarantees they are present in the relevant `__all__` list.

Fixes #3408