fix(python): cannot import nested submodules (#3540)
Nested submodules were not registered in the `__all__` list of their
parent, and so were not visible to import statement such as
`from fully.qualified.submodule.path import Name`.

This changes how submodules get loaded so that they are always declared
and loaded with the context of their immediate parent sub•module, which
guarantees they are present in the relevant `__all__` list.

Fixes #3408



---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0