fix: instantiate perf observer for each iteration (#3585)
Changes to instantiate a `PerformanceObserver` instance for each
iteration of a benchmark run. This prevents the observer from being
killed before all iterations of the wrapped function run due to
environment changes that may cause the observer from falling out of
scope and being GCed. Previously in github actions only a single run's
results were recorded as the observer was killed after the first
iteration, though this wasn't the case when running locally.

Additionally adds the option to record profiles for benchmarks and write
them to a directory. This can be used to compare call stacks of
benchmarks across changes to the code.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0