chore: use fast-deep-equal instead of deep-equal (#3597)
It turns out `fast-deep-equal` is *significantly* faster than
`deep-equal` and appears to produce consistent results. Migrating to
this allows us to yield faster compile times as `deepEqual` is
frequently used when resolving type union references.

Also de-duplicated a couple of calls to `Case.pascal`, which is also
a pretty significant time consumer (but `fast-case` is not iso-functional
with `case`, so this isn't as easy to replace as `deep-equal`).

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0