chore: cache heavy elements in assemblyRelativeSourceFile (#3598)
The readFileSync calls there tend to process large JSON documents, which
is time-consuming. Basic profiling showed this is called repeatedly with
similar directories, so the time spent can be greatly reduced.



---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0