chore(jsii): cache results of case conversions (#3602)
The `Case` library appears on top of the heavy weights when profiling
`jsii` compilation runs on `aws-cdk-lib`, accounting for about `9.6%` of
the overall execution time (`6,712` out of `68,935` ms). This is quite
significant.

Faster libraries performing case conversions exist, however they do not
offer the same level of functinoality that `Case` does, and often
require knowing which casing the string is presently at, when `Case`
performs normalization before conversion (this likely explains why these
conversions are somewhat expensive), which is convenient in our
use-case. It does not seem to be possible to replace `Case` with a
faster implementation, as there does not seem to be any.

Running a survey on the usage of all functions from `Case` being used
during usch a compilation resulted in the following results:

```
Case.camel: 1881755 cached out of 1893235 (99.39%) - 11480 entries / 164.9 hits per entry
Case.pascal: 1890795 cached out of 1909959 (99.00%) - 19164 entries / 99.7 hits per entry
Case.snake: 1882773 cached out of 1895359 (99.34%) - 12586 entries / 150.6 hits per entry
```

Since there is a cache hit rate of `99%` and entries are re-used `100`
times each on average, it appears that adding a cache on those functions
will help alleviate that cost. Indeed, after making this change, `Case`
is no longer present above the `100ms` mark in the list of heavy
weights produced by the profiler.

In order to avoid possibly trading performance for out-of-memory errors,
the caches are keyed off of weak references, so that the caches can be
garbage collected when under memory pressure.



---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0