chore: stop aliasing typescript in dependencies
Aliasing `typescript` to `typescript-3.9` results in occasionally
surprising hoisting behavior, in particular using `npm 18`. In
particular, if two distinct  versions of `typescript` are installed,
aliasing allows both to be hoisted at the top-level, which results in
ambiguity as to which of them gets to have the `node_modules/.bin/tsc`
symlink.

This PR stops aliasing `typescript` to `typescript-3.9` and does the
necessary work to ensure builds continue to operate smoothly, menaing
it replaced TypeScript configuration files for jest with plain ESM
configuration files including typed JSDoc comments (for IDE supprot to
continue working as before), as `jest` otherwise uses `ts-node` to
transform the configuration files, and  `ts-node` uses the "most local"
`typescript` library to perform this parsing (and unfortunately,
`typescript@3.9` does not support the `ES2020` target we are using).