chore: stop aliasing typescript in dependencies (#3623)
Aliasing `typescript` to `typescript-3.9` results in occasionally
surprising hoisting behavior, in particular using `npm 18`. In
particular, if two distinct  versions of `typescript` are installed,
aliasing allows both to be hoisted at the top-level, which results in
ambiguity as to which of them gets to have the `node_modules/.bin/tsc`
symlink.

This PR stops aliasing `typescript` to `typescript-3.9` and does the
necessary work to ensure builds continue to operate smoothly, menaing
it replaced TypeScript configuration files for jest with plain ESM
configuration files including typed JSDoc comments (for IDE supprot to
continue working as before), as `jest` otherwise uses `ts-node` to
transform the configuration files, and  `ts-node` uses the "most local"
`typescript` library to perform this parsing (and unfortunately,
`typescript@3.9` does not support the `ES2020` target we are using).


This also required allowing `jsii.tsc.types` to be specified in the
`package.json` file of jsii projects, as otherwise the TypeScript
compiler attempts to load `@types/*` packages that are not compatible
with `typescript@3.9` (for example, `@synclair/typebox` requires
`typescript@>=4.5`). This setting is anyway generally useful in
complex monorepo situations.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0