chore(kernel): remove unnecessary v8 sandbox (#3657)
Use of the v8 sandbox had already been reduced to the point of being
entirely unnecessary. Finish removing what is left of it, as it also
turns out to make inline debugging a lot easier to perform.



---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0