feat(python): add dynamic type checking (#3660)
Use `typeguard` to perform runtime type checking of arguments passed
into methods (static or instance), setters, and constructors. This
ensures a pythonic error message is produced (and raised as a
`TypeError`), to help developers identify bugs in their code and fix
them.

These checks are disabled when running Python in optimized mode (via
`python3 -O`, which sets `__debug__` to false).

Fixes #3639 

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0