feat(.net): dynmic type checking for union-typed parameters (#3668)
Weaving runtime type checks around union-typed parameters to help
developers receive actionable error messages in case mistakes are made.
The checks are only performed if the
`Amazon.JSII.Runtime.Configuration.RuntimeTypeChecking` configuration
property is `true` (which it is by default).

A pre-processor macro such as `DEBUG` could not be used as the packages
published to NuGet are already built and this does not afford any control
to the end-user.

Fixes #3640



---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0