chore(go): emit one file per type (#3698)
Instead of generating one file will ALL the types in a given
(sub)module, emit one file per type within the (sub)module,
so that the compiler operates on smaller compilation units.
This should reduce the memory pressure the compiler
experiences when dealing with large libraries.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0