fix(dotnet): AnonymousObject faisl runtime type checks
The `AnonymousObject` class is used when a value is passed from JS to
.NET through a union or `any` typed return point, and there is not
sufficient runtime type information on the value to decisively identify
its dynamic type. The `AnonymousObject` class will be converted by the
jsii runtime to any interface type implicitly (even if that is
technically not correct), and it must hence be allowed through runtime
type-checks for interfaces. Essentially, this is a blind spot of the
runtime type checks, and should not cause valid code to break.