fix(dotnet): AnonymousObject fails runtime type checks (#3709)
The `AnonymousObject` class is used when a value is passed from JS to
.NET through a union or `any` typed return point, and there is not
sufficient runtime type information on the value to decisively identify
its dynamic type. The `AnonymousObject` class will be converted by the
jsii runtime to any interface type implicitly (even if that is
technically not correct), and it must hence be allowed through runtime
type-checks for interfaces. Essentially, this is a blind spot of the
runtime type checks, and should not cause valid code to break.



---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0