fix(.net): missing dynamic type checking for collection-nested unions (#3720)
The .NET runtime type checking was not checking type unions that
are nested within a collection (list or map), although this is
necessary since the `is <type>` guard does not allow making any
assumption about the contents of the collection.

This adds the missing checks, and replaces the guard clauses
with a `switch` statement using pattern matching, resulting
in somewhat more elegant code (also with better `null` safety).

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0