fix(python): reference to type does not use alias (#3728)
Type references in a keyword property position may not have used
the aliased import as intended, resulting in a runtime error due
to referring to an undefined symbol.

Additionally, keyword properties referencing structs did not
allow dicts to be used in lieu of a struct instance.



---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0