fix(jsii): missing deprecation warning in collections
When a deprecated member was nested under a collection of values, no
deprecation warning was being generated, despite all the necessary
information being available.

There was a missing case that caused the collection to be validated as
if it was the element type, which is obviously incorrect, but thankfully
did not cause errors as it typically led to verifying `undefined`, which
silently succeeds.

Added the necessary code so that collections are deep-validated as they
should have been.

Fixes #3755