fix(pacmak): .NET bindings fail to compile with error CS8120
If a type union includes several candidates that are related to each
other (A extends B or A implements B), `jsii-pacmak` may generate type
checking clauses in a pattern matching `switch` statement in an order
such that the compiler identifies dead clauses, which is an error in C#.

This adds a provision to NOT emit such a clause so as to not cause the
error. It is worth mentioning that the error cannot be "opted out" of
via a `#pragma` directive like a warning would be, which is unfortunate.

Fixes #3759