chore(spec): improve assembly validation error messages (#3790)
use the ajv error message formatter instead of only outputting the `e.message` as the later fails to include useful information to determine where the actual error is.

Turn the validation test into a snapshot test to draw attention on the error message format.



---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0