fix(python): KeyError in type checks when decorating methods
When methods are decorated by users (e.g: replaced with an alternate
function that delegates back to the original one), type annotations are
not carried over to the new function.

Since type checking code relied on dynamically accessing the checked
function for the purpose of getting type hints,this resulted in
unexpected errors when executing type checking code.

In order to address this, the type checking code now declares a stub
function locally with the relevant type information in order to have a
reliable/stable source of type annotations (these cannot be constructed
dynamically as Python does not expose the necessary constructors).