feat: allow silencing @jsii/check-node warnings
Allows silencing @jsii/check-node warnings when the runtime is not
End-of-Life, so that users who know what they are doing can suppress
the extraneous output.

This is particularly useful for new Node releases that are un-tested,
but can typically be expected to work, or known broken releases that
happen to work for a given use-case that is carefully crafted to avoid
the known issues; but is also possible for deprecated node releases.

Messages about using end-of-life releases are intentionally not
suppressible.

Also updated the message for un-tested and known-broken releases to
better frame the fact that these might work fine, and urge users to
retry using a known supported release before they file a bug report.

Fixes #3817
Fixes #3171