fix(python): parameter names shadow imported modules
In certain cases, a function parameter may shadow an imported module
name, resultin in run-time errors either during type-checking or during
the type-cast that is performed before returning a kernel call's result.

This change adds a test case that covers this particular scenario in
Python, and changes how foreign modules are imported so that an alias
is always generated for those, removin the risk for collisions.

Fixes https://github.com/aws/aws-cdk/issues/22975