fix(python): returning Promise<void> breaks runtime (#3916)
The `EndResponse` struct did not supply a default value for the `result` property, making it required (despite it is typed `Any`, which `None` is valid for). Fix the schema to represent as `Optional[Any] = None` and add a compliance test to validate this scenario.

Note for readers here - async support is very much experimental and only partially imported in most jsii front-end languages. Caution is still advised if you are going to expose async APIs.



---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0