fix(rosetta): submodule-qualified names produce invalid Go, Java
Java and Go do not support transitive access to a submodule's items via
a qualified reference and instead require a dedicated import to be
emitted.

This adds code to analyze use of imported symbols to detect when a
namespace traversal occurs, to inject new imports & replace property
access expressions as necessary.

In order to facilitate this work, these extra elements were done:
- Replaced if-tree in the dispatcher with a switch statement, making the
  dispatch a lot faster, and also a lot nice to run through in a
  debugger
- Emit `var` instead of a type name for variable declarations with an
  initializer in C#, to reduce "invalid" code generated in cases where
  the type name needs some qualification that we are missing
- A couple of bug fixes here and there as the tests discovered them

Fixes #3551