fix(rosetta): go may incorrectly emit _ instead of .
The go transliteration uses a heuristic to determine if a property
access expression is possibly a type name reference, however it failed
to check whether the lead contains a call expression, which would make
it impossible that the overall expression is a type reference.

This fixes the regular expression to address this.