fix(rosetta): go may incorrectly emit _ instead of . (#3985)
The go transliteration uses a heuristic to determine if a property access expression is possibly a type name reference, however it failed to check whether the lead contains a call expression, which would make it impossible that the overall expression is a type reference.

This fixes the regular expression to address this.

Backports a fix from [4.9](https://github.com/aws/jsii-rosetta/pull/7).

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0