fix(dotnet): target netcoreapp3.1 with RollForward (#3987)
Issues where encountered with SDK compatibility when the target
was set to `net6.0` even when the .NET SDK 6.0 or newer was in
use. I suspect this is because of conflicting framework dependency
requirements across various packages.

Re-target `netcoreapp3.1` while setting a roll-forward policy allowing
use of newer major-releases of the SDK when no exact match is found (this
is tested to work correctly for all supported .NET SDK releases by our
automated test suite).