fix(kernel): json deserialization fails on `null` value
The JSON serialization class is a quasi-synonym for `Map<string, any>`,
except it is meant to accept any valid JSON object, including one with
`null` values. The serializer was however interpreting these as
`Map<string, JSON>` when wired as a `$jsii.map` envelope, and this
serialization class does NOT allow for null values.