fix(pacmak): runtime dependency should match pacmak release (#4013)
The code generated by `jsii-pacmak` has a dependency on jsii runtime versions that are expected to match that of `jsii-pacmak` itself. This fixes the .NET and Python generators using the assembly's `jsiiVersion` property, which is a metadata field supposed to only indicate what version of the compiler was used to build the assembly.



---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0