fix(rosetta): flakey trim-cache test
It appears the trim-cache test using tablet compression occasionally
fails in CI/CD. I suspect this is due to the `save` function returning
before the GZip stream has bene closed, resulting in the subsequent read
attempting to consume an incomplete GZip object.

This adds the necessary provisions to ensure the GZip stream is closed
(or failed) before `save` returns.