feat(jsii-pacmak): parsing arguments in non-strict mode
Some targets support extra target-specific arguments so users can
customize the behavior of the build tools. For example, the Java target
allows users to customize the `mvn` CLI behaviour with `--mvn-*` parameters.
However the `jsi--pacmak` CLI is blocking this feature because it is
using strict mode when parsing argument. This change lifts that gate.