fix(pacmak): `_` is not treated as a keyword in Java
Since Java 9, `_` is considered a keyword. Although we are technically
still targeting Java 8 with our source, if we want our source to
be parseable by more modern tools or are considering raising the
threshold at some point, it would be good to have this in place already.

This might technically be considered breaking in case someone is
using `_` as an identifier in a public API, since in newer releases
it will be renamed. But  given that most jsii library authors will be
targeting users that will be running on recent Java versions, and Java 9
is ~6 years old, the chance of that happening in practice is vanishingly
small.