chore: suppress extraneous upgrade prompts (#4092)
`jsii-pacmak` has a dependency on `jsii-rosetta`, which has a dependency on `jsii`. An upgrade prompt will be printed for `jsii` and then `jsii-rosetta`, despite customers have no way to influence the version that is installed as a transitive dependency from `jsii-pacmak`.

This adds suppressions such that:
- `jsii-pacmak` suppresses the warning before loading `jsii-rosetta`
- `jsii-rosetta` suppresses further warnings after having emitted its own, and now emits the warning first thing when it loads (instead of last before)



---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0