feat(go): register exported properties as callbacks
When passing arbitrary structs through an `interface{}` type, the JS
value had no properties or methods, leading to it having an empty object
JSON representation.

In order to improve the situation, discover exported properties of such
structs, and register them as property callbacks, more closely matching
the behavior of JS in the same configuration.

The callbacks use the same naming transformation as the standard JSON
serialization mechanism in go: using the `json` tag if present with the
same semantics as `encoding/json`, and fall back to the raw field name
otherwise (with no case conversion applied).

Related: cdk8s-team/cdk8s#1326