fix(jsii-pacmak): improve struct performance in Java
When a struct is returned to Java, it gets returned by reference. The
Java struct proxy then proceeds to read all fields from the struct, even
if they are never used by the Java program.

Make this behavior on-demand with a cache: the value is read if it is
requested, but will never be read more than once (this is fine as
structs are immutable).

If we decide to pass the struct back into jsii, we can avoid having
to read all the fields at that point to build a fresh struct if we
hold on to the struct reference and pass that back in.