feat(jsii): allow specifying a deprecation message regardless of stability (#4145)
Previously, `jsii` required the package's stability to be `deprecated` if there cas a `deprecated` message configured. This is however problematic when a `stable` package reaches End-of-Support and the maintainer wants to signal this in npmjs.com.

Instead of failing, only log a warning to possibly raise awareness on an unintended configuration.

Additionally - emit the `Development Status :: 7 - Inactive` trove classifier on packages with a `deprecated` message, even if their stability is not `deprecated`.