chore(runtime): correct return types for completeCallback and processRequest (#4143)
The return type for `completeCallback` was incorrectly listed as `void`, which confused me for a bit. This PR corrects it.

Changes here:

* I've changed the return type of `completeCallback` to `api.CompleteRequest['result']` (was `void`) as this is correct. Technically, this resolves to `any` - but I thought `api.CompleteRequest['result']` was more explicit about what is actually happening.
* I templated the callback argument to `processRequest` so that the return type is correct there, too.
* I've given both `handleCallback` and explicit return types. This is personal preference - let me know if you'd prefer the functions left with implicit return types.

As an aside, I was surprised the old version compiled, but it turns out you're allowed to return an `any` from a function marked as returning `void`. Makes sense as you can assign `any` to anything, but it was a bit surprising.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0